+++ ming/odm/odmsession.py
@@ -1,5 +1,6 @@
 from collections import defaultdict
 
+from pymongo.collection import ReturnDocument
 from pymongo.database import Database
 
 from ming.session import Session
@@ -188,11 +189,10 @@ class ODMSession:
         _call_hook(self, 'cursor_created', odm_cursor, 'find', cls, *args, **kwargs)
         return odm_cursor
 
-    # FIXME: delete
-    def find_and_modify(self, cls, *args, **kwargs):
+    def __find_and_modify(self, cls, operation: str, *args, **kwargs):
         """Finds and updates ``cls``.
 
-        Arguments are the same as :meth:`pymongo.collection.Collection.find_and_modify`.
+        Arguments are the same as :meth:`pymongo.collection.Collection.find_one_and_update`.
 
         If the session has ``autoflush`` option, the session
         if flushed before performing the query.
@@ -204,12 +204,22 @@ class ODMSession:
         if self.autoflush:
             self.flush()
         m = mapper(cls)
-        obj = self.impl.find_and_modify(m.collection, *args, **kwargs)
+        fn = getattr(self.impl, operation)
+        obj = fn(m.collection, *args, **kwargs)
         if obj is None: return None
         cursor = ODMCursor(self, cls, iter([ obj ]), refresh=True, decorate=decorate)
         result = cursor.first()
         state(result).status = ObjectState.clean
         return result
+
+    def find_one_and_update(self, cls, *args, **kwargs):
+        return self.__find_and_modify(cls, 'find_one_and_update', *args, **kwargs)
+
+    def find_one_and_replace(self, cls, *args, **kwargs):
+        return self.__find_and_modify(cls, 'find_one_and_replace', *args, **kwargs)
+
+    def find_one_and_delete(self, cls, *args, **kwargs):
+        return self.__find_and_modify(cls, 'find_one_and_delete', *args, **kwargs)
 
     @_with_hooks('remove')
     def remove(self, cls, *args, **kwargs):